Removed empty test cases. Added integration test package that will be used to test EF and repos in test database.
