+++ StorageTests/DbContextTests.cs
@@ -5,6 +5,10 @@ using Storage;
 
 namespace StorageTests
 {
+
+    /// <summary>
+    /// This class has been used for internal test purposes to see if could connect to database. 
+    /// </summary>
     [TestClass]
     public class DbContextTests
     {
@@ -17,7 +21,11 @@ namespace StorageTests
             _dbContext = new AutoSysDbModel();
         }
 
+        /// <summary>
+        /// This test method is used to see if it is possible to connect to the database. 
+        /// </summary>
         [TestMethod]
+        [Ignore]
         public void CheckConnection()
         {
             bool result; 
@@ -36,8 +44,11 @@ namespace StorageTests
             Assert.AreEqual(true, result);
         }
 
-
+        /// <summary>
+        /// Checks if it is possible to create the database. 
+        /// </summary>
         [TestMethod]
+        [Ignore]
         public void CreateDatabase()
         {
             try
