--- StorageTests/RepositoryTests/TaskRepositoryTests.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace StorageTests
-{
-    [TestClass]
-    public class TaskRepositoryTests
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
-    }
-}
