--- StorageTests/RepositoryTests/TeamRepositoryTests.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace StorageTests
-{
-    [TestClass]
-    public class TeamRepositoryTests
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
-    }
-}
