+++ StorageTests/RepositoryUnitTests/TeamRepositoryTests.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace StorageTests.RepositoryTests
+{
+    [TestClass]
+    public class TeamRepositoryTests
+    {
+        
+    }
+
+}
