+++ StorageTests/RepositoryUnitTests/UserRepositoryTests.cs
@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
-using Storage;
 using Storage.Models;
-using Storage.Repository;
-using Storage.Repository.Interface;
 using StorageTests.Utility;
 
 namespace StorageTests.RepositoryTests
@@ -21,13 +16,13 @@ namespace StorageTests.RepositoryTests
     /// 
     /// </summary>
     [TestClass]
-    public class AsyncUserRepositoryTests
+    public class UserRepositoryTests
     {
 
         private IList<StoredUser> _data;
-        private Mock<IAutoSysContext> _context; // Use IUserContext instead of concrete AutoSysDbModel
+        private Mock<Utility.IUserContext> _context; // Use IUserContext instead of concrete AutoSysDbModel
         private Mock<DbSet<StoredUser>> _mockSet;
-        private AsyncDbRepository<StoredUser> _repository;
+        private UserRepositoryStub _repository;
 
         /// <summary>
         /// This method sets up data used to mock a collection of users in a DbContext used by the UserRepository. 
@@ -40,14 +35,13 @@ namespace StorageTests.RepositoryTests
             {
                 new StoredUser {Id = 1, Name = "William Parker", MetaData = "Researcher"},
                 new StoredUser {Id = 2, Name = "Trudy Jones", MetaData = "Researcher"}
-            };
+            }; 
 
-            _mockSet = MockUtility.CreateAsyncMockDbSet(_data, u => u.Id);
+            _mockSet = MockUtility.CreateMockDbSet(_data, u => u.Id);
 
-            var mockContext = new Mock<IAutoSysContext>();
+            var mockContext = new Mock<IUserContext>();
             mockContext.Setup(s => s.Users).Returns(_mockSet.Object);
-            mockContext.Setup(s => s.Set<StoredUser>()).Returns(_mockSet.Object);
-            mockContext.Setup(s => s.SaveChangesAsync()).Returns(Task.Run(() =>
+            mockContext.Setup(s => s.SaveChanges()).Callback(() =>
             {
                 // Increment user ids automatically based on existing ids in mock data 
                 var max = _data.Max(u => u.Id);
@@ -55,27 +49,10 @@ namespace StorageTests.RepositoryTests
                 {
                     user.Id = ++max;
                 }
-                return 0;
-            }));
+            });
 
             _context = mockContext;
-            _repository = new AsyncDbRepository<StoredUser>(_context.Object);
-        }
-
-        [TestMethod, ExpectedException(typeof(ArgumentNullException))]
-        public async Task Create_NullException()
-        {
-            await _repository.Create(null);
-        }
-
-        [TestMethod]
-        public async Task Create_SaveChanged_IsCalled()
-        {
-            var user = new StoredUser();
-            var id = await _repository.Create(user);
-
-            _context.Verify(c => c.SaveChangesAsync(), Times.Once);
-
+            _repository = new UserRepositoryStub(_context.Object);
         }
 
         /// <summary>
@@ -86,47 +63,47 @@ namespace StorageTests.RepositoryTests
         /// </summary>
         /// <returns></returns>
         [TestMethod]
-        public async Task Create_SavesUserInContext()
+        public void Create_SavesUserInContext()
         {
             // Arrange 
             var mockSet = new Mock<DbSet<StoredUser>>();
-            var mockContext = new Mock<IAutoSysContext>();
+            var mockContext = new Mock<IUserContext>();
             mockContext.Setup(m => m.Users).Returns(mockSet.Object);
 
             // Act 
-            var service = new AsyncDbRepository<StoredUser>(mockContext.Object);
+            var service = new UserRepositoryStub(mockContext.Object);
             //var service = new DbRepositoryStub<StoredUser>(mockContext.Object);
-            var id = await service.Create(new StoredUser());
+            int id = service.Create(new StoredUser());
 
             // Assert 
             mockSet.Verify(m => m.Add(It.IsAny<StoredUser>()), Times.Once());
-            mockContext.Verify(m => m.SaveChangesAsync(), Times.Once());
+            mockContext.Verify(m => m.SaveChanges(), Times.Once());
 
         }
 
         [TestMethod]
         public void Create_SaveChanges_IsCalled()
         {
-            var user = new StoredUser();
-            var id = _repository.Create(user);
-            _context.Verify(r => r.SaveChangesAsync(), Times.Once);
+                var user = new StoredUser();
+                var id = _repository.Create(user);
+                _context.Verify(r => r.SaveChanges(), Times.Once);
         }
 
         [TestMethod]
-        public async Task Create_ReturnsId_NewId()
+        public void Create_ReturnsId_NewId()
         {
             // Arrange 
             var mockSet = new Mock<DbSet<StoredUser>>();
-            var mockContext = new Mock<IAutoSysContext>();
+            var mockContext = new Mock<IUserContext>();
             mockContext.Setup(m => m.Users).Returns(mockSet.Object);
             var user = new StoredUser { Name = "Steven", MetaData = "Validator" };
 
             // Act 
             //var service = new DbRepositoryStub<StoredUser>(mockContext.Object);
-            var service = new AsyncDbRepository<StoredUser>(mockContext.Object);
-            var id = await service.Create(user);
+            var service = new UserRepositoryStub(mockContext.Object);
+            var id = service.Create(user);
 
-            Assert.AreEqual(0, user.Id); // True for EF but not for interface 
+            Assert.AreEqual(0, user.Id);
         }
 
         [TestMethod]
@@ -134,14 +111,14 @@ namespace StorageTests.RepositoryTests
         {
             // Arrange 
             var mockSet = new Mock<DbSet<StoredUser>>();
-            var mockContext = new Mock<IAutoSysContext>();
+            var mockContext = new Mock<IUserContext>();
             mockContext.Setup(m => m.Users).Returns(mockSet.Object);
             var user = new StoredUser { Name = "Steven", MetaData = "Validator" };
-            var secondUser = new StoredUser { Name = "William", MetaData = "Researcher" };
+            var secondUser = new StoredUser {Name = "William", MetaData = "Researcher"};
 
             // Act 
             //var service = new DbRepositoryStub<StoredUser>(mockContext.Object);
-            var service = new AsyncDbRepository<StoredUser>(mockContext.Object);
+            var service = new UserRepositoryStub(mockContext.Object);
             var id = service.Create(user);
             var secondId = service.Create(secondUser);
 
@@ -150,64 +127,62 @@ namespace StorageTests.RepositoryTests
         }
 
         [TestMethod]
-        public async Task Update_SaveChanges_IsCalled()
+        public void Update_SaveChanges_IsCalled()
         {
-            var user = new StoredUser { Id = 1 };
-            await _repository.Update(user);
-            _context.Verify(r => r.SaveChangesAsync(), Times.Once);
+                var user = new StoredUser { Id = 1 };
+                _repository.UpdateIfExists(user);
+                _context.Verify(r => r.SaveChanges(), Times.Once);
         }
 
         [TestMethod]
-        public async Task Update_NewName()
+        public void Update_NewName()
         {
             // Arrange 
             var user = new StoredUser { Id = 1, Name = "User", MetaData = "Validator" };
             var researcher = _data[0];
 
             // Act 
-            await _repository.Update(user);
+            _repository.UpdateIfExists(user);
 
-
-            _context.Verify(c => c.SetModified(user), Times.Once);
             // Act 
             Assert.AreEqual("User", researcher.Name);
         }
 
         [TestMethod]
-        public async Task Update_NewMetaData()
+        public void Update_NewMetaData()
         {
             // Arrange 
             var researcher = _data[0];
             var user = new StoredUser { Id = 1, MetaData = "Validator" };
 
             // Act 
-            await _repository.Update(user);
+            _repository.UpdateIfExists(user);
 
             // Assert
             Assert.AreEqual("Validator", researcher.MetaData);
         }
 
         [TestMethod]
-        public async Task Delete_RemovesUser()
+        public void Delete_RemovesUser()
         {
             // Arrange 
             var user = _data[0];
 
             // Act 
-            var id = await _repository.Delete(user.Id);
+            _repository.DeleteIfExists(user);
 
             // Assert
             Assert.IsFalse(_data.Any(u => u.Id == 1));
         }
 
         [TestMethod]
-        public async Task Delete_SaveChanges_IsCalled()
+        public void Delete_SaveChanges_IsCalled()
         {
             // Arrange
             var user = _data[0];
 
             // Act
-            await _repository.Delete(user.Id);
+            _repository.DeleteIfExists(user);
 
             // Assert
             _context.Verify(repo => repo.SaveChanges(), Times.Once);
@@ -215,9 +190,9 @@ namespace StorageTests.RepositoryTests
 
 
         [TestMethod]
-        public async Task GetById()
+        public void GetById()
         {
-            var secondUser = await _repository.Read(1);
+            var secondUser = _repository.Read(1);
             Assert.AreEqual("William Parker", secondUser.Name);
         }
